+++ packages/expo-router/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Fix pinned `react-navigation` dependencies ([#43456](https://github.com/expo/expo/pull/43456) by [@kitten](https://github.com/kitten))
 - Fix zoom transition to prefetched routes ([#43852](https://github.com/expo/expo/pull/43852) by [@Ubax](https://github.com/Ubax))
+- Fix `Stack.Protected` guard not applying to index routes.
 
 ### 💡 Others